+++ app/Controller/CategoriesController.php
@@ -61,17 +61,22 @@ class CategoriesController extends AppController {
  * @return void
  */
 	public function add() {
-		if ($this->request->is('post')) {
-			$this->Category->create();
-			if ($this->Category->save($this->request->data)) {
-				$this->Session->setFlash(__('The category has been saved.'));
-				return $this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('The category could not be saved. Please, try again.'));
+		if($this->Auth->user('is_author')) {
+			if ($this->request->is('post')) {
+				$this->Category->create();
+				if ($this->Category->save($this->request->data)) {
+					$this->Session->setFlash(__('The category has been saved.'));
+					return $this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('The category could not be saved. Please, try again.'));
+				}
 			}
+			$articles = $this->Category->Article->find('list');
+			$this->set(compact('articles'));
+		} else {
+			$this->Session->setFlash(__('You Cannot Access this Area')); 
+			$this->redirect(array('action' => 'index')); 
 		}
-		$articles = $this->Category->Article->find('list');
-		$this->set(compact('articles'));
 	}
 
 /**
@@ -82,22 +87,27 @@ class CategoriesController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
-		if (!$this->Category->exists($id)) {
-			throw new NotFoundException(__('Invalid category'));
-		}
-		if ($this->request->is(array('post', 'put'))) {
-			if ($this->Category->save($this->request->data)) {
-				$this->Session->setFlash(__('The category has been saved.'));
-				return $this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('The category could not be saved. Please, try again.'));
+		if($this->Auth->user('is_author')) {
+			if (!$this->Category->exists($id)) {
+				throw new NotFoundException(__('Invalid category'));
 			}
+			if ($this->request->is(array('post', 'put'))) {
+				if ($this->Category->save($this->request->data)) {
+					$this->Session->setFlash(__('The category has been saved.'));
+					return $this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('The category could not be saved. Please, try again.'));
+				}
+			} else {
+				$options = array('conditions' => array('Category.' . $this->Category->primaryKey => $id));
+				$this->request->data = $this->Category->find('first', $options);
+			}
+			$articles = $this->Category->Article->find('list');
+			$this->set(compact('articles'));
 		} else {
-			$options = array('conditions' => array('Category.' . $this->Category->primaryKey => $id));
-			$this->request->data = $this->Category->find('first', $options);
+			$this->Session->setFlash(__('You Cannot Access this Area')); 
+			$this->redirect(array('action' => 'index')); 
 		}
-		$articles = $this->Category->Article->find('list');
-		$this->set(compact('articles'));
 	}
 
 /**
@@ -108,17 +118,22 @@ class CategoriesController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		$this->Category->id = $id;
-		if (!$this->Category->exists()) {
-			throw new NotFoundException(__('Invalid category'));
-		}
-		$this->request->onlyAllow('post', 'delete');
-		if ($this->Category->delete()) {
-			$this->Session->setFlash(__('The category has been deleted.'));
+		if($this->Auth->user('is_author')) {
+			$this->Category->id = $id;
+			if (!$this->Category->exists()) {
+				throw new NotFoundException(__('Invalid category'));
+			}
+			$this->request->onlyAllow('post', 'delete');
+			if ($this->Category->delete()) {
+				$this->Session->setFlash(__('The category has been deleted.'));
+			} else {
+				$this->Session->setFlash(__('The category could not be deleted. Please, try again.'));
+			}
+			return $this->redirect(array('action' => 'index'));
 		} else {
-			$this->Session->setFlash(__('The category could not be deleted. Please, try again.'));
+			$this->Session->setFlash(__('You Cannot Access this Area')); 
+			$this->redirect(array('action' => 'index')); 
 		}
-		return $this->redirect(array('action' => 'index'));
 	}
 
 /**
